requirements.txt
transformers==4.30.2
datasets==2.12.0
peft==0.4.0
trl==0.7.1
accelerate==0.21.0
bitsandbytes==0.41.0
torch==2.0.1
evaluate==0.4.0
rouge_score==0.1.2
nltk==3.8.1
gradio==3.40.1
wandb==0.15.8
numpy==1.24.3
pandas==2.0.3
tqdm==4.65.0
python-dotenv==1.0.0
requests==2.31.0
PyYAML==6.0
scipy==1.10.1